backend/requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
httpx==0.28.1
python-dotenv==1.2.1
langchain==0.3.27
langchain-openai==0.3.35
langgraph==0.6.11
